Prepare for generating Rust code alongside OCaml generation
