+++ charon/Cargo.toml
@@ -48,8 +48,8 @@ name = "charon-driver"
 path = "src/bin/charon-driver/main.rs"
 
 [[bin]]
-name = "generate-ml"
-path = "src/bin/generate-ml/main.rs"
+name = "generate-asts"
+path = "src/bin/generate-asts/main.rs"
 
 [[test]]
 name = "ui"
