+++ charon/src/bin/generate-asts/generate_ml/templates/Expressions.ml
@@ -1,11 +1,11 @@
 (** WARNING: this file is partially auto-generated. Do not edit `src/Expressions.ml`
-    by hand. Edit `templates/Expressions.ml` instead, or improve the code
+    by hand. Edit `generate_ml/templates/Expressions.ml` instead, or improve the code
     generation tool so avoid the need for hand-writing things.
 
-    `templates/Expressions.ml` contains the manual definitions and some `(*
+    `generate_ml/templates/Expressions.ml` contains the manual definitions and some `(*
     __REPLACEn__ *)` comments. These comments are replaced by auto-generated
-    definitions by running `make generate-ml` in the crate root. The
-    code-generation code is in `charon/src/bin/generate-ml`.
+    definitions by running `make generate-asts` in the crate root. The
+    code-generation code is in `charon/src/bin/generate-asts`.
  *)
 open Identifiers
 open Generated_Types
